fix(rp): SDK 2.x pio_program init + type-limits + sign-compare hygiene (#2727) (#2740)
Three RP2040/RP2350 warning families fire on every CI build for those
boards. None are immediate bugs individually, but together they signal
the RP common layer needs a hygiene pass before pico-sdk 2.x lands as
the default for RP2350.

## H1. pio_program missing initializer for pio_version / used_gpio_ranges

pico-sdk 2.x (RP2350) added `pio_version` and `used_gpio_ranges` to
`struct pio_program`. FastLED's four `pio_program` literals omit them,
triggering -Wmissing-field-initializers and (more dangerously) leaving
the new fields zero-initialized by C++ aggregate rules. On RP2350
"no GPIO range claimed" could become a runtime failure the moment the
SDK starts validating it.

Fix: explicitly initialize both new fields to 0 under
`#if defined(PICO_SDK_VERSION_MAJOR) && PICO_SDK_VERSION_MAJOR >= 2`
guards in all 4 literals (pio_gen.h, spi_hw_2_rp.cpp.hpp,
spi_hw_4_rp.cpp.hpp, spi_hw_8_rp.cpp.hpp). Preserves current behavior,
silences the warning, and documents the field-init choice in one place
to revisit when RP2350-SDK ergonomics solidify.

## H2. -Wtype-limits "always true" on `mData[1-3]Pin >= 0`

In `spi_hw_4_rp.cpp.hpp` the data-lane pin presence checks read
`if (mData1Pin >= 0 && mData2Pin >= 0 && mData3Pin >= 0)` — but the
fields are u8, so the comparison is always true. The `SpiHw4::Config`
source field is `i8` with `-1` as the "unused lane" sentinel, so the
intent was clearly "lane is configured" — currently broken: the
storage truncates `-1` to `0xFF` and every lane is treated as active.

Fix: change `mData1Pin / mData2Pin / mData3Pin` from `u8` to `i8` so
`-1` round-trips. Update constructor initializers from `0` to `-1` to
match the Config defaults. `mData0Pin` and `mClockPin` stay u8 (always
required, never sentinel-valued). Lane-presence checks now work as
intended.

## H3. -Wsign-compare in pin_rp.hpp (`int pin` vs unsigned NUM_BANK0_GPIOS)

Four loop-bound checks in pin_rp.hpp compare `int pin` against
`NUM_BANK0_GPIOS` (unsigned, from pico-sdk). Cast the bound to int
inline — the cast is exact since NUM_BANK0_GPIOS fits in int on both
RP2040 and RP2350.

Closes #2727.

+++ src/platforms/arm/rp/pin_rp.hpp
@@ -115,7 +115,7 @@ inline u16 analogRead(int pin) FL_NOEXCEPT {
 /// PWM frequency is approximately 244 Hz (system clock / 65536).
 inline void analogWrite(int pin, u16 val) FL_NOEXCEPT {
     // Check if pin supports PWM
-    if (pin >= NUM_BANK0_GPIOS) {
+    if (pin >= static_cast<int>(NUM_BANK0_GPIOS)) {  // #2727 -Wsign-compare
         return;  // Invalid pin for PWM
     }
 
@@ -146,7 +146,7 @@ inline void analogWrite(int pin, u16 val) FL_NOEXCEPT {
 /// PWM frequency is approximately 1.9 Hz @ 125 MHz (system clock / 65536).
 inline void setPwm16(int pin, u16 val) FL_NOEXCEPT {
     // Check if pin supports PWM
-    if (pin >= NUM_BANK0_GPIOS) {
+    if (pin >= static_cast<int>(NUM_BANK0_GPIOS)) {  // #2727 -Wsign-compare
         return;  // Invalid pin for PWM
     }
 
@@ -228,7 +228,7 @@ inline bool needsPwmIsrFallback(int /*pin*/, u32 frequency_hz) {
 /// The PWM output is enabled when analogWrite() or setPwm16() is called.
 inline int setPwmFrequencyNative(int pin, u32 frequency_hz) FL_NOEXCEPT {
     // Validate pin
-    if (pin < 0 || pin >= NUM_BANK0_GPIOS) {
+    if (pin < 0 || pin >= static_cast<int>(NUM_BANK0_GPIOS)) {  // #2727 -Wsign-compare
         return -1;
     }
 
@@ -284,7 +284,7 @@ inline int setPwmFrequencyNative(int pin, u32 frequency_hz) FL_NOEXCEPT {
 /// @param pin GPIO pin number (0-29)
 /// @return Configured frequency in Hz, or 0 if not configured
 inline u32 getPwmFrequencyNative(int pin) FL_NOEXCEPT {
-    if (pin < 0 || pin >= NUM_BANK0_GPIOS) {
+    if (pin < 0 || pin >= static_cast<int>(NUM_BANK0_GPIOS)) {  // #2727 -Wsign-compare
         return 0;
     }
     Rp2040PwmFreq* table = rp_pwm_freq_table();
